Added support for dist config. Moved common http/tftp code to separate file. Fixed syntax error in dhcp-hook.py.
