+++ dhcp-hook.py
@@ -26,6 +26,6 @@ if sys.argv[1] == "commit":
     sql = "SELECT short_name FROM network WHERE ipv4_gateway_txt = ?"
     networkname = cursor.execute(sql, (swRelay, )).fetchone()[0]
     db.set('networkname-{}'.format(swIp), networkname)
-  if "Juniper" not in :swClient
+  if "Juniper" not in swClient:
     # We don't need any SNMP or base config for Juniper.
     os.system("/scripts/swboot/configure " + swIp + " &")
