+++ dhcpd.conf
@@ -20,7 +20,7 @@ on commit {
   if agentType = "0" {
     set swName = pick-first-value(binary-to-ascii(16, 8, ":", substring(option agent.circuit-id, 4, 2)), "None");
   } else {
-    set swName = pick-first-value(substring(option agent.circuit-id, 2, 10), "None");
+    set swName = pick-first-value(substring(option agent.circuit-id, 2, 50), "None");
   }
   set swMac = binary-to-ascii(16, 8, ":", substring(hardware, 1, 6));
   set swIp = binary-to-ascii(10, 8, ".", leased-address);
