+++ swhttpd.py
@@ -1,15 +1,13 @@
 #!/usr/bin/env python
 import sys, logging
-import redis
 import syslog
 import socket
-import re
-import tempfile
 import SimpleHTTPServer
 import SocketServer
 import time
 
 import config
+import swcommon
 
 def log(*args):
   print time.strftime("%Y-%m-%d %H:%M:%S") + ':', ' '.join(args)
@@ -21,45 +19,25 @@ def error(*args):
 
 class swbootHttpHandler(SimpleHTTPServer.SimpleHTTPRequestHandler):
   def do_GET(self):
-    db = redis.Redis()
-    switch = db.get(self.client_address[0])
-    model = db.get('client-{}'.format(self.client_address[0]))
-    if switch == None or model == None:
+    # self.path is the path of the requested file.
+    file_handle = swcommon.select_file(self.path.lstrip("/"), self.client_address[0])
+
+    if file_handle == None:
       log("Switch not found:", self.client_address[0])
       self.send_error(404, "File not found")
       return None
-    if self.path == "/juniper-confg":
-      log("Generating Juniper config for",
-          self.client_address[0], "name =", switch)
-      f = tempfile.TemporaryFile()
-      f.write(config.generate(switch, model))
-      content_length = f.tell()
-      f.seek(0)
 
-      self.send_response(200)
-      self.send_header("Content-type", "application/octet-stream")
-      self.send_header("Content-Length", content_length)
-      self.end_headers()
-      self.copyfile(f, self.wfile)
-      log("Config sent to", self.client_address[0], "name =", switch)
+    # Go to the end of the file to get the length of it.
+    file_handle.seek(0, 2)
+    content_length = file_handle.tell()
+    file_handle.seek(0)
 
-      f.close()
-      return
-    elif self.path == "/juniper.tgz":
-      log("Sending JunOS file", config.models[model]['image'], "to",
-        self.client_address[0], "name =", switch)
-      if (model in config.models) and ('image' in config.models[model]):
-        # All good! Overwrite the requested file path and send our own.
-        self.path = config.models[model]['image']
-        f = self.send_head()
-        if f:
-          self.copyfile(f, self.wfile)
-          log("Sent JunOS to", self.client_address[0], "name =", switch)
-          f.close()
-    else:
-      log("Unknown file:", self.path)
-      self.send_error(404, "File not found")
-  
+    self.send_response(200)
+    self.send_header("Content-type", "application/octet-stream")
+    self.send_header("Content-Length", content_length)
+    self.end_headers()
+    self.copyfile(file_handle, self.wfile)
+
   # We write our own logs.
   def log_request(self, code='-', size='-'):
     pass
@@ -68,6 +46,7 @@ class swbootHttpHandler(SimpleHTTPServer.SimpleHTTPRequestHandler):
 
 class swbootTCPServer(SocketServer.ForkingTCPServer):
   def server_bind(self):
+    self.socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
     self.socket.bind(self.server_address)
 
 log("swhttpd started")
